% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_nmme}
\alias{download_nmme}
\title{NMME URLs}
\usage{
download_nmme(x, overwrite = FALSE, ...)
}
\arguments{
\item{x}{nmme.}

\item{overwrite}{Logical. Whether or not files should be downloaded
again if they already exist.}

\item{...}{Additional arguments to download.file.}
}
\value{
NULL.
}
\description{
This function returns a character vector of NMME URLs.
}
\examples{
\dontrun{
x <- nmme()
download_nmme(x[1,1,1])
}
}
