% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter.nmme}
\alias{filter.nmme}
\title{Filter nmme object.}
\usage{
\method{filter}{nmme}(.data, ...)
}
\arguments{
\item{.data}{nmme object.}

\item{...}{
  Arguments passed on to \code{\link[=filter_init_time]{filter_init_time}}, \code{\link[=filter_member]{filter_member}}, \code{\link[=filter_lead_time]{filter_lead_time}}, \code{\link[=filter_point]{filter_point}}, \code{\link[=filter_region]{filter_region}}
  \describe{
    \item{\code{start_init_time}}{zoo::yearmon. Start initialization time.}
    \item{\code{end_init_time}}{zoo::yearmon. End initialization time.}
    \item{\code{members}}{Integer.}
    \item{\code{lead_times}}{Numeric.}
    \item{\code{xcoord}}{Numeric.}
    \item{\code{ycoord}}{Numeric.}
    \item{\code{xmin}}{Numeric.}
    \item{\code{xmax}}{Numeric.}
    \item{\code{ymin}}{Numeric.}
    \item{\code{ymax}}{Numeric.}
  }}
}
\description{
Filter nmme object.
}
