% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter_init_time}
\alias{filter_init_time}
\title{Filter by initialization time}
\usage{
filter_init_time(.data, start_init_time, end_init_time)
}
\arguments{
\item{.data}{nmme object.}

\item{start_init_time}{zoo::yearmon. Start initialization time.}

\item{end_init_time}{zoo::yearmon. End initialization time.}
}
\value{
List.
}
\description{
Filter by initialization time
}
\examples{
\dontrun{
print("Hello, world!")
}
}
