% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter_lead_time}
\alias{filter_lead_time}
\title{Filter by lead times}
\usage{
filter_lead_time(.data, lead_times)
}
\arguments{
\item{.data}{nmme object.}

\item{lead_times}{Numeric.}
}
\value{
List.
}
\description{
Filter by lead times
}
\examples{
\dontrun{
print("Hello, world!")
}
}
