% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter_member}
\alias{filter_member}
\title{Filter by members}
\usage{
filter_member(.data, members)
}
\arguments{
\item{.data}{nmme object.}

\item{members}{Integer.}
}
\value{
List.
}
\description{
Filter by members
}
\examples{
\dontrun{
print("Hello, world!")
}
}
