% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter_point}
\alias{filter_point}
\title{Filter by point.}
\usage{
filter_point(.data, xcoord, ycoord)
}
\arguments{
\item{.data}{nmme object.}

\item{xcoord}{Numeric.}

\item{ycoord}{Numeric.}
}
\value{
List.
}
\description{
Filter by point.
}
\examples{
\dontrun{
print("Hello, world!")
}
}
