% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter_region}
\alias{filter_region}
\title{Filter by region.}
\usage{
filter_region(.data, xmin, xmax, ymin, ymax)
}
\arguments{
\item{.data}{nmme object.}

\item{xmin}{Numeric.}

\item{xmax}{Numeric.}

\item{ymin}{Numeric.}

\item{ymax}{Numeric.}
}
\value{
List.
}
\description{
Filter by region.
}
\examples{
\dontrun{
print("Hello, world!")
}
}
