% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iri_config.R
\name{iri_config}
\alias{iri_config}
\title{IRI configuration}
\usage{
iri_config(...)
}
\arguments{
\item{...}{Additional arguments (None implemented).}
}
\value{
A precipitation object.
}
\description{
Return a list with IRI...
}
\examples{
\dontrun{
sum(1:10)
}
}
