% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmme.R
\name{nmme}
\alias{nmme}
\title{NMME}
\usage{
nmme(
  model = "GFDL-SPEAR",
  simulation_type = "hindcast",
  dataset = "monthly",
  variable = "prec",
  config = nmme_config(),
  ...
)
}
\arguments{
\item{model}{Character.}

\item{simulation_type}{Character.}

\item{dataset}{Character.}

\item{config}{List.}

\item{...}{Additional arguments.}
}
\value{
List.
}
\description{
This function returns an NMME object.
}
\examples{
\dontrun{
urls = nmme()
download_nmme(urls[1])
}
}
