% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{nmme_index}
\alias{nmme_index}
\title{Index URLs}
\usage{
nmme_index(x, dir, ...)
}
\arguments{
\item{x}{nmme object.}

\item{dir}{character.}

\item{...}{Additional arguments.}
}
\description{
Construct URLs from \code{nmme} object.
}
