% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetting.R
\name{nmme_subset}
\alias{nmme_subset}
\title{Subset nmme object.}
\usage{
nmme_subset(x, ...)
}
\arguments{
\item{x}{nmme object.}

\item{...}{
  Arguments passed on to \code{\link[=nmme_subset_init_times]{nmme_subset_init_times}}, \code{\link[=nmme_subset_members]{nmme_subset_members}}, \code{\link[=nmme_subset_lead_times]{nmme_subset_lead_times}}, \code{\link[=nmme_subset_point]{nmme_subset_point}}, \code{\link[=nmme_subset_region]{nmme_subset_region}}
  \describe{
    \item{\code{start_init_time}}{zoo::yearmon. Start initialization time.}
    \item{\code{end_init_time}}{zoo::yearmon. End initialization time.}
    \item{\code{config}}{TODO.}
    \item{\code{members}}{Integer.}
    \item{\code{lead_times}}{Numeric.}
    \item{\code{xcoord}}{Numeric.}
    \item{\code{ycoord}}{Numeric.}
    \item{\code{xmin}}{Numeric.}
    \item{\code{xmax}}{Numeric.}
    \item{\code{ymin}}{Numeric.}
    \item{\code{ymax}}{Numeric.}
  }}
}
\description{
Subset nmme object.
}
