% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetting.R
\name{nmme_subset_region}
\alias{nmme_subset_region}
\title{Subset by region.}
\usage{
nmme_subset_region(xmin, xmax, ymin, ymax)
}
\arguments{
\item{xmin}{Numeric.}

\item{xmax}{Numeric.}

\item{ymin}{Numeric.}

\item{ymax}{Numeric.}
}
\value{
List.
}
\description{
Subset by region.
}
\examples{
\dontrun{
print("Hello, world!")
}
}
