% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.nmme}
\alias{print.nmme}
\title{Print nmme object}
\usage{
\method{print}{nmme}(x, ...)
}
\arguments{
\item{x}{nmme object.}

\item{...}{Additional arguments (ignored).}
}
\description{
Print nmme object
}
