% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\name{subset}
\alias{subset}
\title{Subset nmme object.}
\usage{
subset(x, ...)
}
\arguments{
\item{x}{nmme object.}

\item{...}{
  Arguments passed on to \code{\link[=subset_init_times]{subset_init_times}}, \code{\link[=subset_members]{subset_members}}, \code{\link[=subset_lead_times]{subset_lead_times}}, \code{\link[=subset_point]{subset_point}}, \code{\link[=subset_region]{subset_region}}
  \describe{
    \item{\code{start_init_time}}{zoo::yearmon. Start initialization time.}
    \item{\code{end_init_time}}{zoo::yearmon. End initialization time.}
    \item{\code{config}}{TODO.}
    \item{\code{members}}{Integer.}
    \item{\code{lead_times}}{Numeric.}
    \item{\code{xcoord}}{Numeric.}
    \item{\code{ycoord}}{Numeric.}
    \item{\code{xmin}}{Numeric.}
    \item{\code{xmax}}{Numeric.}
    \item{\code{ymin}}{Numeric.}
    \item{\code{ymax}}{Numeric.}
  }}
}
\description{
Subset nmme object.
}
