% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\name{subset_init_times}
\alias{subset_init_times}
\title{Subset by initialization time}
\usage{
subset_init_times(start_init_time, end_init_time, config)
}
\arguments{
\item{start_init_time}{zoo::yearmon. Start initialization time.}

\item{end_init_time}{zoo::yearmon. End initialization time.}

\item{config}{TODO.}
}
\value{
List.
}
\description{
Subset by initialization time
}
\examples{
\dontrun{
print("Hello, world!")
}
}
