% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\name{subset_lead_times}
\alias{subset_lead_times}
\title{Subset by lead times}
\usage{
subset_lead_times(lead_times, config)
}
\arguments{
\item{lead_times}{Numeric.}

\item{config}{TODO}
}
\value{
List.
}
\description{
Subset by lead times
}
\examples{
\dontrun{
print("Hello, world!")
}
}
