% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\name{subset_members}
\alias{subset_members}
\title{Subset by members}
\usage{
subset_members(members, config)
}
\arguments{
\item{members}{Integer.}

\item{config}{TODO}
}
\value{
List.
}
\description{
Subset by members
}
\examples{
\dontrun{
print("Hello, world!")
}
}
