% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\name{subset_point}
\alias{subset_point}
\title{Subset by point.}
\usage{
subset_point(xcoord, ycoord)
}
\arguments{
\item{xcoord}{Numeric.}

\item{ycoord}{Numeric.}
}
\value{
List.
}
\description{
Subset by point.
}
\examples{
\dontrun{
print("Hello, world!")
}
}
