% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_nmme.R
\name{download_nmme}
\alias{download_nmme}
\title{NMME URLs}
\usage{
download_nmme(x, destdir = ".", overwrite = FALSE, ...)
}
\arguments{
\item{destdir}{Character. Destination directory, which is created
if it doesn't already exist.}

\item{overwrite}{Logical. Whether or not files should be downloaded
again if they already exist.}

\item{...}{Additional arguments to download.file.}

\item{urls}{List of character pairs. The first value of each pair
should be the URL, the second value should be the destination
filename.}
}
\value{
NULL.
}
\description{
This function returns a character vector of NMME URLs.
}
\examples{
\dontrun{
urls = name_urls()
download_nmme(urls[1])
}
}
